Move `\SMW\PropertySpecificationLookup` to `\SMW\Property\SpecificationLookup`, refs 3801
